// SGUEES-SPA/src/app/pages/General/gen-distrito/models/gen-distrito.ts
export interface GenDistrito {
	CORR_EMPRESA: number;
	CORR_PAIS: number;
	CORR_DEPTO: number;
	CORR_MUNICIPIO: number;
	CORR_DISTRITO: number;
	NOMBRE_DISTRITO: string;
	USUARIO_CREA: string;
	ESTACION_CREA: string;
	FECHA_CREA: Date;
	USUARIO_ACTU: string;
	ESTACION_ACTU: string;
	FECHA_ACTU: Date;
	NOMBRE_PAIS: string;
	NOMBRE_DEPTO: string;
	NOMBRE_MUNICIPIO: string;
}